Better support "mark unread" and retain all updates
Closes #4

This change maintains two "read" timestamps, the most recent as well as
one previous. This allows the logic to marking a pull request as unread
to revert back to the previous timestamp. As well, this change retains
all activity, even if the user has marked that activity as read. This
means that when marking a ticket as unread there is no chance of losing
context. Finally, it is retaining even changes the user made themselves.
This is more a hopeful change with the idea that it can be used
elsewhere possibly to paint a timeline of the ticket.